# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Weekly Commitments API
  description: |
    API for the Weekly Commitments module – plan creation, lifecycle management,
    reconciliation, manager review, RCDO integration, and AI-assisted suggestions.
  version: 1.0.0
  contact:
    name: ST6 Engineering
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /admin/ai-usage:
    get:
      operationId: getAiUsage
      summary: Get AI feature usage metrics (admin only)
      description: |
        Returns suggestion acceptance/defer/decline rates from
        ai_suggestion_feedback and in-process cache hit/miss statistics.
        Requires the ADMIN role.
      tags: [Admin]
      parameters:
        - name: weeks
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 26
            default: 8
          description: Rolling-window size in weeks (1–26)
      responses:
        '200':
          description: AI usage metrics returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageMetrics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AiUsageMetrics:
      type: object
      required:
        - weeks
        - windowStart
        - windowEnd
        - totalFeedbackCount
        - acceptedCount
        - deferredCount
        - declinedCount
        - acceptanceRate
        - cacheHits
        - cacheMisses
        - cacheHitRate
        - approximateTokensSpent
        - approximateTokensSaved
      description: AI feature usage metrics over a rolling window.
      properties:
        weeks:
          type: integer
          description: Number of weeks in the rolling window
        windowStart:
          type: string
          format: date
          description: Start of the rolling window
        windowEnd:
          type: string
          format: date
          description: End of the rolling window
        totalFeedbackCount:
          type: integer
          format: int64
          description: Total suggestion feedback records in the window
        acceptedCount:
          type: integer
          format: int64
          description: Count of ACCEPT actions
        deferredCount:
          type: integer
          format: int64
          description: Count of DEFER actions
        declinedCount:
          type: integer
          format: int64
          description: Count of DECLINE actions
        acceptanceRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: acceptedCount / totalFeedbackCount (0 if no feedback)
        cacheHits:
          type: integer
          format: int64
          description: |
            Cumulative AI cache hits attributed to this org since the last service restart.
        cacheMisses:
          type: integer
          format: int64
          description: Cumulative AI cache misses attributed to this org since the last service restart
        cacheHitRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: cacheHits / (cacheHits + cacheMisses) (0 if no cache activity)
        approximateTokensSpent:
          type: integer
          format: int64
          description: |
            Coarse estimate of AI tokens spent for this org since the last service restart.
            Computed as cacheMisses × 1000 tokens/request.
        approximateTokensSaved:
          type: integer
          format: int64
          description: |
            Coarse estimate of AI tokens avoided by cache hits for this org since the last
            service restart. Computed as cacheHits × 1000 tokens/request.
...
    ApiError:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Valid JWT but insufficient access
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...